refactor: sections moved to section folder cards moved to cards folder
