+++ src/components/homepage/cards/AccountsCard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import clsx from 'clsx';
 import gsap from 'gsap';
-import Card from '../../Card';
-import LineChart from '../../../charts/LineChart';
+import Card from './Card';
+import LineChart from '../../charts/LineChart';
 import useInterval from '@/hooks/useInterval';
 import '@/lib/swapText';
