+++ src/components/homepage/cards/BanksCard.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import clsx from 'clsx';
 import gsap from 'gsap';
-import Card from '../../Card';
-import PieChart from '../../../charts/PieChart';
-import BarChart from '../../../charts/BarChart';
+import Card from './Card';
+import PieChart from '../../charts/PieChart';
+import BarChart from '../../charts/BarChart';
 import useInterval from '@/hooks/useInterval';
 import '@/lib/swapText';
