+++ src/components/homepage/cards/DemoCard.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import clsx from 'clsx'
-import Card from '../../Card';
+import Card from './Card';
 import Button from '@/components/buttons/Button';
 import Ping from '@/components/Ping';
