+++ src/components/homepage/cards/StatisticsCard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import clsx from 'clsx';
 import gsap from 'gsap';
-import Card from '../../Card';
-import BarChart from '../../../charts/BarChart';
+import Card from './Card';
+import BarChart from '../../charts/BarChart';
 import useInterval from '@/hooks/useInterval';
 import '@/lib/swapText';
