+++ src/components/homepage/sections/Instructions.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import clsx from 'clsx';
-import Card from './Card';
+import Card from '../cards/Card';
 import { VscAccount } from 'react-icons/vsc';
 import Image from 'next/image';
-import Button from '../buttons/Button';
-import ButtonLink from '../links/ButtonLink';
+import Button from '../../buttons/Button';
+import ButtonLink from '../../links/ButtonLink';
 import { AiOutlineLineChart } from 'react-icons/ai';
-import Ping from '../Ping';
+import Ping from '../../Ping';
 
 const InstructionsSection = () => {
   return (
